+++ CHANGELOG.md
@@ -29,6 +29,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- cpp: Clean up unused `reset_joint_properties` function
+- cpp: Remove joint properties from Bullet interface
 - utils: Handle directional pad (D-pad) inputs in `Joystick` class
 
 ### Removed